Finally resolved!!!!
In this commit include :
Adding background music, my own twist.
Using default music only for main menu.
The most important is possible to set asteroid spawning on the 50th enemies defeated and the advance enemy will show on the 75th enemies destroyed.
Just to add difficulties to the game on the spesific time or indicator.
Make sure to check the code on SpawnManager script.
Due to some error and crashed i separate the coruotine on sperata methode dedicated spesifically to call the coroutine. Just to be safe.